requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
bson==0.5.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-cors-headers==4.3.1
django-import-export==3.3.8
django-rest-passwordreset==1.4.1
django-user-agents==0.4.0
django-utils-six==2.0
djangorestframework==3.15.1
docopt==0.6.2
et-xmlfile==1.1.0
html5lib==1.1
idna==3.7
lxml==5.2.1
MarkupPy==1.14
#mysql-client==0.0.1
num2words==0.5.13
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
pillow==10.2.0
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pypdf==4.0.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
text-unidecode==1.3
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ua-parser==0.18.0
uritools==4.0.2
urllib3==2.2.0
user-agents==2.2.0
webencodings==0.5.1
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0
